Update/fix Omnibot.Irc.route_msg/2 to spawn a new task for each module, instead of a single task for all modules
Previously, if a message was supposed to be handled by 3 modules, they
would all be handled, synchronously, in the same process. If any of them
crashed, it would affect any other modules that needed to be processed
ahead of it.

Now, a new task is spun up for each module, so module handlers are now
indpendent.

Signed-off-by: Alek Ratzloff <alekratz@gmail.com>